+++ tests/calculators/vasp/test_vasp.py
@@ -54,7 +54,7 @@ def test_presets():
     assert calc.exp_params["ediff"] == 1e-5
     assert calc.float_params["encut"] == 450
 
-    calc = Vasp(atoms, xc="scan", preset="MPScanSet")
+    calc = Vasp(atoms, xc="scan", preset="MPR2SCANSet")
     assert calc.xc.lower() == "scan"
     assert calc.string_params["algo"] == "all"
     assert calc.exp_params["ediff"] == 1e-5
@@ -160,7 +160,7 @@ def test_magmoms():
 
     atoms = bulk("Cu") * (2, 2, 1)
     atoms[-1].symbol = "Fe"
-    calc = Vasp(atoms, preset="MPScanSet")
+    calc = Vasp(atoms, preset="MPR2SCANSet")
     atoms.calc = calc
     assert atoms.get_initial_magnetic_moments().tolist() == [1.0] * (len(atoms) - 1) + [
         5.0
@@ -600,7 +600,7 @@ def test_setups():
     assert calc.parameters["setups"]["Cu"] == ""
 
     atoms = bulk("Cu")
-    calc = Vasp(atoms, preset="MPScanSet")
+    calc = Vasp(atoms, preset="MPR2SCANSet")
     assert calc.parameters["setups"]["Cu"] == "_pv"
 
     atoms = bulk("Cu")
@@ -620,7 +620,7 @@ def test_setups():
     assert calc.parameters["setups"]["Cu"] == ""
 
     atoms = bulk("Cu")
-    calc = Vasp(atoms, setups="minimal", preset="MPScanSet")
+    calc = Vasp(atoms, setups="minimal", preset="MPR2SCANSet")
     assert isinstance(calc.parameters["setups"], str)
     assert calc.parameters["setups"] == "minimal"